--- state/multiwatcher/w.go
@@ -1,76 +1,0 @@
-package multiwatcher
-
-import (
-	"bytes"
-	"encoding/json"
-	"fmt"
-
-	"github.com/juju/juju"
-)
-
-// Delta holds details of a change to the environment.
-type Delta struct {
-	// If Removed is true, the entity has been removed;
-	// otherwise it has been created or changed.
-	Removed bool
-	// Entity holds data about the entity that has changed.
-	Entity EntityInfo
-}
-
-// MarshalJSON implements json.Marshaler.
-func (d *Delta) MarshalJSON() ([]byte, error) {
-	b, err := json.Marshal(d.Entity)
-	if err != nil {
-		return nil, err
-	}
-	var buf bytes.Buffer
-	buf.WriteByte('[')
-	c := "change"
-	if d.Removed {
-		c = "remove"
-	}
-	fmt.Fprintf(&buf, "%q,%q,", d.Entity.EntityId().Kind, c)
-	buf.Write(b)
-	buf.WriteByte(']')
-	return buf.Bytes(), nil
-}
-
-// UnmarshalJSON implements json.Unmarshaler.
-func (d *Delta) UnmarshalJSON(data []byte) error {
-	var elements []json.RawMessage
-	if err := json.Unmarshal(data, &elements); err != nil {
-		return err
-	}
-	if len(elements) != 3 {
-		return fmt.Errorf(
-			"Expected 3 elements in top-level of JSON but got %d",
-			len(elements))
-	}
-	var entityKind, operation string
-	if err := json.Unmarshal(elements[0], &entityKind); err != nil {
-		return err
-	}
-	if err := json.Unmarshal(elements[1], &operation); err != nil {
-		return err
-	}
-	if operation == "remove" {
-		d.Removed = true
-	} else if operation != "change" {
-		return fmt.Errorf("Unexpected operation %q", operation)
-	}
-	switch entityKind {
-	case "machine":
-		d.Entity = new(juju.MachineInfo)
-	case "service":
-		d.Entity = new(juju.ServiceInfo)
-	case "unit":
-		d.Entity = new(juju.UnitInfo)
-	case "relation":
-		d.Entity = new(juju.RelationInfo)
-	case "annotation":
-		d.Entity = new(juju.AnnotationInfo)
-	default:
-		return fmt.Errorf("Unexpected entity name %q", entityKind)
-	}
-	return json.Unmarshal(elements[2], &d.Entity)
-}
